feat: 🎸 sync chain info
✅ Closes: #4
